% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeCog.R
\name{analyzeCog}
\alias{analyzeCog}
\title{analyze cogs}
\usage{
analyzeCog(cog, ...)
}
\value{
the output of the supplied cog
}
\description{
applies set methods to analyze a supplied cog
}
\details{
backend methods to calculate cogs for a \code{gear} object
}
