% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeGear.R
\name{analyzeGear}
\alias{analyzeGear}
\title{analyzeGear}
\usage{
analyzeGear(GDS, gear = NULL, ...)
}
\arguments{
\item{GDS}{\code{character} contating the path to a GDS file 
or an already imported \code{SeqVarGDSClass}}

\item{gear}{\code{gear} an object created with \code{makeGear()} contating the desired analyses. \cr
Analyses that can be carried out are (see \code{?makeCog() for details}): \cr
Genetic Diversity
Admixture
Output Loci 
Output Trees}
}
\value{
the supplied object with cogs replaced by results
}
\description{
analyzes the object class \code{gear}
}
\details{
analyszes a supplied GDS using a \code{gear} object containing the desired 
cogs.
}
