% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCodonIndex.R
\name{buildCodonDB}
\alias{buildCodonDB}
\title{build a codon sqlLite DB}
\usage{
buildCodonDB(genome, exons, sqlDir = NULL, ...)
}
\arguments{
\item{genome}{\code{character} or \code{DNAStringSet} \cr
Import fasta reference genome from file by passsing a \code{character} or provide a preloaded (using \code{Biostrings::readDNAStringSet()}) \code{DNAStringSet}}

\item{exons}{\code{GrangesList} \cr
Generated using \code{getFeatures()} by passing \code{"gene:cds"} to the  \code{feature} argument.}

\item{sqlDir}{\code{character} set to sql output to run in lower memory mode.}

\item{position}{\code{character}
Options are \code{"all"}, default, or any combination of \code{c("first", "second", "third")}. \cr
Output will contain only specified codon positions.}

\item{fourFoldCodon}{\code{character}
 Options are \code{NULL}, \code{"only"} or \code{"remove"}. \cr
Determines what to do with 4-fold degenerate sites. \cr}
}
\value{
An object of class \code{GRanges} containing codon ranges for positions specified.
}
\description{
Index genome according to codon positions in coding sequence. \cr
This can be useful when downstream analysis or output of alignements should only be applied to certain codons. \cr
For example calculating genetic distance on only 4-fold sites or outputting fasta files based on codon position.
}
\examples{

}
