% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog-class.R
\docType{class}
\name{cog.args-class}
\alias{cog.args-class}
\alias{cog-class,cog.args}
\alias{cog.diversityFAST-class}
\alias{cog-class,cog.diversityFAST}
\alias{cog.diversityFULL-class}
\alias{cog-class,cog.diversityFULL}
\alias{cog.admixture-class}
\alias{cog-class,cog.admixture}
\alias{cog.outputLoci-class}
\alias{cog-class,cog.outputLoci}
\alias{cog.outputTrees-class}
\alias{cog-class,cog.outputTrees}
\alias{cog.NULL-class}
\alias{cog-class,cog.NULL}
\title{Cog classes}
\value{
An object of class cog.X
}
\description{
Cog classes
}
\details{
cog classes to fit into gear object slots
}
