% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineBUSCOgff.R
\name{combineBUSCOgff}
\alias{combineBUSCOgff}
\title{Combines gffs output by BUSCO}
\usage{
combineBUSCOgff(inDir, outDir, outName, geneList = NULL)
}
\arguments{
\item{inDir}{directory to input single gffs from. gffs should be in *geneName*.gff nomenclature}

\item{outDir}{directory to output combined gff}

\item{geneList}{only import genes from this vector/list}
}
\value{
A gff contatining all genes in geneList
}
\description{
Combines gffs from BUSCO annotation of a genome
}
\details{
Authour: Chris Ward
}
\examples{



}
