% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmax.R
\name{dmax}
\alias{dmax}
\title{Calculates maximum genetic distance}
\usage{
dmax(distMat, popList, pairs, ploidy)
}
\arguments{
\item{distMat}{hamming distance matrix calculated using the hidden function \code{genoDist}}

\item{popList}{List of populations made from \code{pops} dataframe provided by users}

\item{pairs}{list of populaiton pairs generated in \code{getDiversityStats}}

\item{ploidy}{\code{numeric} number of chromosomes}
}
\value{
A \code{dataframe} maximum genetic distance between samples
}
\description{
Calculates maximum population genetic distance from hamming distance
}
\details{
Calculates maximum population genetic distance from hamming distance
}
