% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmin.R
\name{dmin}
\alias{dmin}
\title{Calculates minimum genetic distance}
\usage{
dmin(distMat, popList, pairs, ploidy)
}
\arguments{
\item{distMat}{hamming distance matrix calculated using the hidden function \code{genoDist}}

\item{popList}{List of populations made from \code{pops} dataframe provided by users}

\item{pairs}{list of populaiton pairs generated in \code{getDiversityStats}}

\item{ploidy}{\code{numeric} number of chromosomes}
}
\value{
A \code{dataframe} mean minimum genetic distance between samples
}
\description{
Calculates minimum population genetic distance from hamming distance
}
\details{
Authours: Chris Ward
Calculates minimum population genetic distance from hamming distance
}
