% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dxyAF.R
\name{dxyAF}
\alias{dxyAF}
\title{#' Calculates geentic distance between populations using allele frequency at biallelic SNPS
#'
#' @description Used in the calculation of diversity statistics FAST
#'
#' @details Authours: Chris Ward
#' Calculates geentic distance between populations using allele frequency at biallelic SNPS
#'
#'
#' @param AF allele frequency
#' @param pop populaitons dataframe
#'
#' @return A \code{matrix} of genetic distance between populaitons
#'
#' @useDynLib geaR
#'
#'
#' @rdname dxyAF
#' @export}
\usage{
dxyAF(AF)
}
\description{
#' Calculates geentic distance between populations using allele frequency at biallelic SNPS
#'
#' @description Used in the calculation of diversity statistics FAST
#'
#' @details Authours: Chris Ward
#' Calculates geentic distance between populations using allele frequency at biallelic SNPS
#'
#'
#' @param AF allele frequency
#' @param pop populaitons dataframe
#'
#' @return A \code{matrix} of genetic distance between populaitons
#'
#' @useDynLib geaR
#'
#'
#' @rdname dxyAF
#' @export
}
