% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate model parameters}
\usage{
estimate(object, ...)
}
\arguments{
\item{object}{A model object produced by the
\code{\link[gear]{geolm}} function.}

\item{...}{Currently unimplemented}
}
\description{
\code{estimate} estimates the parameters of the specified
model. The function is written to automatically adapt
based on the class of \code{object}.  Currently, the
\code{estimate} function performs maximum likelihood
estimation for objects produced by the
\code{\link[gear]{geolm}} function.
}
\examples{
set.seed(10)
n = 100
}
\seealso{
\code{\link[gear]{estimate.geolm_cmodStd}},
}
\author{
Joshua French
}
