% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCodonDB.R
\name{filterCodonDB}
\alias{filterCodonDB}
\title{filter codon position and type}
\usage{
filterCodonDB(DB, position, degeneracy)
}
\arguments{
\item{DB}{\code{GRangeList} or \code{character} vector containing the path to a SQL database}

\item{position}{\code{character} any combination of \code{c("first", "second", "third")}.}

\item{degeneracy}{\code{character} this will select positions based on degeneracy can be one of \code{c(0, 2)}. \cr 
This will overwrite the \code{position} argument. \cr 
For four fold degenerate sites use \code{valicate4fold()}.}
}
\value{
An object of class \code{GRanges} containing codon ranges for positions specified.
}
\description{
Use a codon database to filter codons based on position and degeneracy
}
\examples{

}
