% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gear-class.R
\docType{class}
\name{gear-class}
\alias{gear-class}
\title{The gear Object Class}
\value{
An object of class geaR
}
\description{
The gear Object Class
}
\details{
This object contains \code{cog} objects to allow for multiple analyses to
be carried out on the same loci
}
