% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoDist.R
\name{genoDist}
\alias{genoDist}
\title{Calculates the hamming distance between alleles}
\usage{
genoDist(genoMat, pairwiseDeletion)
}
\arguments{
\item{genoMat}{A \code{matrix} \cr
Allele genotypes for each individual}

\item{pairwiseDeletion}{\code{logical}
If \code{TRUE} missing data will be removed from distance calculations in a paiwise manner.}
}
\value{
A \code{matrix} of hamming distance between individuals
}
\description{
Used in the calculation of diversity statistics
}
\details{
Authours: Chris Ward
Calculates the hamming distance using matrix multiplication
}
