% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractionFunctions.R
\name{getAdmixFromGear}
\alias{getAdmixFromGear}
\title{getAdmixFromGear}
\usage{
getAdmixFromGear(gear)
}
\arguments{
\item{gear}{\code{gear} object user has analyzed}
}
\value{
dataframe of results
}
\description{
Extract admixture metrics from a gear object
}
