% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAF.R
\name{getAF}
\alias{getAF}
\title{Calculates population based allele frequency}
\usage{
getAF(GDS, locus, minSites, pops, refAllele, counts = FALSE)
}
\arguments{
\item{GDS}{allele frequency}

\item{locus}{Granges locus to act on}

\item{minSites}{minimum sites within a window to act on}

\item{pop}{populations dataframe}
}
\value{
A \code{tibble} of population allele frequency for biallelic SNPs
}
\description{
Calculates population based allele frequency
}
\details{
Authours: Chris Ward
Calculates population based allele frequency
}
