% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCodonIndex.R
\docType{methods}
\name{getCodonIndex}
\alias{getCodonIndex}
\alias{getCodonIndex,CodonIndex-method}
\title{Index genome according to provided exons}
\usage{
getCodonIndex(genome, exons)

\S4method{getCodonIndex}{DNAStringSet}(genome, exons)
}
\arguments{
\item{genome}{genome}

\item{exons}{GrangeList exons}
}
\value{
An object of \code{CodonIndex}
}
\description{
Author: Christopher Ward Index genome according to provided exons
}
\examples{

}
