% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDiversityStats.R
\name{getDiversityStats}
\alias{getDiversityStats}
\title{Calculate diversity statistics}
\usage{
getDiversityStats(GDS, loci, minSites = 0.5, nCores = 1, pops = NULL,
  stats = "all", ploidy = 2, pairwiseDeletion, ...)
}
\arguments{
\item{GDS}{\code{character} or \code{SeqVarGDSClass} \cr
File path pointing to \code{GDS} file or \code{SeqVarGDSClass} object imported usign \code{SeqArray::seqOpen()}.}

\item{loci}{\code{GRanges} or \code{GRangesList}. \cr
Loci to import genotypes for.}

\item{minSites}{\code{numeric} \cr
Minimum number of sites as a proportion of loci length. Default 0.5 i.e. 50 percent}

\item{nCores}{\code{numeric} \cr
Number of cores to run in parallel}

\item{pops}{\code{data_frame} \cr
Containing Sample ID and poplations. Default is to treat individuals as populations.}

\item{stats}{\code{character} \cr
Vector containing all diversity stats to calculate. Default \code{c("all")}. \cr
Currently supports \code{c("pi", "dxy", "da", "dmin", "dmax", "Fst")} \cr
\cr
"pi": Neis nucleotide diversity \cr
"dxy": Neis absolute genetic distance \cr
"da": Nei's ancesteral distance \cr
"dmin": minimum hamming distance between samples \cr
"dmax": maximum hamming distance between samples \cr
"Fst": Nei (1982) \eqn{\gamma}st which estimates Fst}

\item{ploidy}{\code{numeric} \cr
Ploidy of samples}

\item{pairwiseDeletion}{\code{logical} \cr
If \code{TRUE} missing data will be removed from distance calculations in a paiwise manner.}
}
\value{
A \code{data_frame} of selected Diversity statistics
}
\description{
Imports genotypes and calculates diversity formula
}
\details{
Authour: Chris Ward \cr
Calculates Diversity statstics along \code{GRanges} or a \code{GRangesList} to import genotypes
}
