% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractionFunctions.R
\name{getDivFromGear}
\alias{getDivFromGear}
\title{getDivFromGear}
\usage{
getDivFromGear(gear)
}
\arguments{
\item{gear}{\code{gear} object user has analyzed}
}
\value{
dataframe of results
}
\description{
Extract Diversity metrics from a gear object
}
