% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDensity.R, R/getGenes.R
\name{ThreeTipDistance}
\alias{ThreeTipDistance}
\alias{getFeatures}
\title{Makes Density plots}
\usage{
ThreeTipDistance(data, I1, I2, O)

getFeatures(gffName, contigMD, feature = "gene", nCores = 1,
  longestIsoform = FALSE, which = NULL)
}
\arguments{
\item{gffName}{path to gff to extract features from}

\item{contigMD}{data frame with contig metadata. Contains two columns: \code{ID} with scaffold names, \code{length} with scaffold length}

\item{feature}{\code{list} or \code{GRangesList}. Loci to import genotypes for. c("gene", "gene:exon", "gene:cds", "pseudogene", "lncRNA", "intergenic")}

\item{nCores}{\code{numeric} number of cores to run in parallel}

\item{longestIsoform}{\code{logical} only effects feature = "gene:exon". by default select the first entry for each gene. If \code{TRUE} will select the longest isoform for that gene. If there is a biotype field in the GFF it will only deal with those labeled as 'protein coding.' Selecting the longest isoform may cause be problematic if you do not have the biotype.}

\item{which}{t}

\item{minSites}{\code{numeric} minimum number of sites as a proportion of loci length. Default 0.5 (ie 50 percent)}

\item{gffName}{path to gff to extract features from}

\item{contigMD}{data frame with contig metadata. Contains two columns: \code{ID} with scaffold names, \code{length} with scaffold length}

\item{feature}{\code{list} or \code{GRangesList}. Loci to import genotypes for. c("gene", "gene:exon", "gene:cds", "pseudogene", "lncRNA", "intergenic")}

\item{minSites}{\code{numeric} minimum number of sites as a proportion of loci length. Default 0.5 (ie 50 percent)}

\item{nCores}{\code{numeric} number of cores to run in parallel}

\item{longestIsoform}{\code{logical} only effects feature = "gene:exon". by default select the first entry for each gene. If \code{TRUE} will select the longest isoform for that gene. If there is a biotype field in the GFF it will only deal with those labeled as 'protein coding.' Selecting the longest isoform may cause be problematic if you do not have the biotype.}
}
\value{
A \code{list} of gene GRanges

A \code{list} of gene GRanges
}
\description{
Plots stats into Density plots

Creates feature loci
}
\details{
Authour: Chris Ward

Authour: Chris Ward
}
\examples{


}
