% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenotyes.R
\name{getGenotypes}
\alias{getGenotypes}
\title{Import genotypes from GDS}
\usage{
getGenotypes(GDS, locus, minSites = 0.5, nCores = 1, nuc = FALSE)
}
\arguments{
\item{GDS}{\code{GDS} object with variant data to import genotypes from}

\item{minSites}{\code{numeric} minimum number of sites as a proportion of loci length}

\item{loci}{Loci to import genotypes for}

\item{ncores}{\code{numeric} number of cores to run in parallel}
}
\value{
A \code{list} of \code{GRanges}
}
\description{
Imports genotypes across loci into matrix
}
\details{
Authours: Chris Ward & Alastair Luddington
generates tiled (no step size) and sliding (step size) ranges for analysis with other functions.
}
\examples{



}
