% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPositionFromDB.R
\name{getPositionFromDB}
\alias{getPositionFromDB}
\title{getPositionFromDB}
\usage{
getPositionFromDB(DB, positions)
}
\arguments{
\item{DB}{\code{GRange or character}. Either the path to a SQL database or GRange object 
containing codon information.}

\item{positions}{\code{character} vector containing any of code{c("first", 
"second", "third")}}
}
\value{
a GRange object containing specified codon positions.
}
\description{
extracts a partition scheme from a codon DB
}
\details{
Accepts both GRange and SQL codon databases generated using \code{buildCodonDB}
}
