% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gmin.R
\name{Gmin}
\alias{Gmin}
\title{Gmin}
\usage{
Gmin(distMat, popList, pairs, ploidy)
}
\arguments{
\item{distMat}{hamming distance matrix calculated using the hidden function \code{genoDist}}

\item{popList}{List of populations made from \code{pops} dataframe provided by users}

\item{pairs}{list of populaiton pairs generated in \code{getDiversityStats}}

\item{ploidy}{\code{numeric} number of chromosomes}
}
\value{
A \code{dataframe} containing Gmin
}
\description{
Calculates Gmin from Geneva (2015)
}
\details{
Authours: Chris Ward \cr
calculates Gmin from ...
}
