% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCog.R
\name{makeCog}
\alias{makeCog}
\title{Make cogs}
\usage{
makeCog(analysisType = "args", ...)
}
\arguments{
\item{analysisType}{\code{character}. Analysis type: \cr}

\item{...}{See examples or the vignette for what arguments to pass in \code{...} \cr
\code{"args"}: Generic arguments for an analysis \cr
\code{"diversityFULL"}: Arguments spcific to the Diversity module \cr

\itemize{
Currently supports: \cr
\item pi: Neis nucleotide diversity \cr
\item dxy: Neis absolute genetic distance \cr
\item da: Nei's ancesteral distance \cr
\item dmin: minimum hamming distance between samples \cr
\item dmax: maximum hamming distance between samples \cr
\item Fst: Nei (1982) \eqn{\gamma}st which estimates Fst \cr
\item RNDmin: minimum Relative node depth
\item federRND: Relative Node Depth using Feder (2005)}
\code{"admixture"}: Arguments spcific to the Admixture module \cr
\code{"outputLoci"}:  Arguments spcific to the Output Loci module \cr
\code{"outputTrees"}: Arguments spcific to the Output Trees module \cr}
}
\value{
Output a \code{cog} object to
}
\description{
Builds the cog to pass to \code{makeGear}
}
\details{
backend methods to build cogs for a \code{gear} object  \cr
Specify the \code{analysisType} and the corresponding argument set
}
\examples{
\dontrun{

# Argument cog
## Necessary arguments are:
## ploidy: The ploidy called in the genotyping process.
## nCores: Number of cores to run analysis over.
## minSites: The proportion of sites within the window with a genotype called. 
## pairwiseDeletion: This will remove missing data in a pairwise manner i.e. will not count missing data as informative.
## removeIndels: Should indels be removed from the analysis?
arg <- makeCog(analysisType = "args", ploidy = 2, nCores = 4, minSites = 0.2, pairwiseDeletion = TRUE, removeIndels = TRUE)

# Genetic Diversity cog
## stats: What statistics should be run? 
diversity <- makeCog(analysisType = "diversityFULL", stats = "all")

# Admixture cog
## threePop: structure for f3 stat 
## fourPop: structure for f4 and fd stat 
admix <- makeCog(analysisType = "admixture", threePop = list(c("P1", "P2", "O")),
                 fourPop = list(c("P1", "P2", "P3", "O")))
                 
# Output Loci cog
## outputDirectory: The directory to output fasta files to for downstream analysis
## alleles: Output "seperate" or "collapsed" genotypes
outloci <-  makeCog(analysisType = "outputLoci", outputDirectory = "~/path/to/dir/", alleles = "seperate")

# Output Trees cog
## outputDirectory: The directory to output fasta files to for downstream analysis
## alleles: Output "seperate" or "collapsed" genotypes
outTrees <-  makeCog(analysisType = "outputTrees", outputDirectory = "~/path/to/dir/", alleles = "seperate")
}

}
