% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gear-class.R
\name{makeGear}
\alias{makeGear}
\title{make gear object}
\usage{
makeGear(loci, populations, outgroup = vector("character", 1), cogs = list())
}
\arguments{
\item{loci}{\code{GRangesList} containing genomic loci to work on}

\item{populations}{\code{data.frame} containing two columns \code{Sample} and
\code{Population}}

\item{outgroup}{\code{character} the population in the \code{populations}
argument that corresponds to the outgroup}

\item{cogs}{\code{list} a list of cog objects for the analysis}
}
\description{
make a gear object from cogs
}
\details{
This object contains \code{cog} objects to allow for multiple analyses to
be carried out on the same loci
}
