% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R, R/mle.geolmStd.R
\name{mle}
\alias{mle}
\title{Finds maximum likelihood estimates of model parameters
for a geostatistical model}
\usage{
mle(
  object,
  reml = FALSE,
  est = "e",
  lower = NULL,
  upper = NULL,
  method = "nlminb",
  itnmax = NULL,
  control = list(),
  ...
)

mle(
  object,
  reml = FALSE,
  est = "e",
  lower = NULL,
  upper = NULL,
  method = "nlminb",
  itnmax = NULL,
  control = list(),
  ...
)
}
\arguments{
\item{object}{A geostatistical linear model object
producted by the \code{geolm} function.}

\item{reml}{A logical value indicating whether standard
maximum likelihood estimation should be performed
(\code{reml = FALSE}).  If \code{reml = TRUE}, then
restricted maximum likelihood is performed.  Defaul is
\code{FALSE}.}

\item{est}{A character vector indicator whether the error
variance (\code{est="e"}) or finescale variance
(\code{est = "f"}) should be estimated.  The other
component of the nugget variance is held constant, and
in the case of a \code{geolmStd} object, is set to 0.}

\item{lower}{A vector of 2 or 3 specifying the lowerbound
of parameter values.  See Details.}

\item{upper}{lower A vector of 2 or 3 specifying the
lowerbound of parameter values.  See Details.}

\item{method}{The optimization method.  Default is
\code{"nlminb"}, with \code{"L-BFGS-B"} being another
acceptable choice.  See \code{\link[optimx]{optimx}}
for details.}

\item{itnmax}{An integer indicating the maximum number of
iterations to allow for the optimization prodedure.}

\item{control}{A list of control parameters passed
internally to \code{\link[optimx]{optimx}}.  See
\code{\link[optimx]{optimx}} for details.}

\item{...}{Currently unimplemented.}
}
\description{
\code{mle} estimates the parameters of a geostatistical
linear model.

\code{mle} estimates the parameters of a geostatistical
linear model. The \code{mle} function will be deprecated
in the future. Please update your code to use the
estimate function.
}
\details{
In the case of a \code{geolmStd} \code{object}, the
likelihood has been concentrated so that only the range
parameter \code{r} and a scale parameter \code{lambda =
nugget/psill} need to be estimated.

If \code{object} is a \code{geolmStd}, then \code{lower}
is of length 2 if the covariance model of \code{cmod} is
not \code{matern} or \code{amatern}.  Otherwise, it
should be of length 3.  The first parameter is related to
the range parameter \code{r}, the second to the scale
parameter \code{lambda}, and the third to \code{par3}, if
applicable.  If \code{lower = NULL}, then the lower
bounds are 0.001, 0, and 0.1, respectively.  A similar
pattern holds for \code{upper}, with the default being
\code{3 * max(d)}, where \code{d} is the matrix of
distances between coordinates, \code{5}, and \code{2.5}.

The \code{kkt} argument in the \code{control} list is set
to be \code{FALSE}.
}
\examples{
set.seed(10)
n = 100
set.seed(10)
n = 100
}
\seealso{
\code{\link[gear]{estimate}},
}
\author{
Joshua French

Joshua French
}
