% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neisda.R
\name{neisDa}
\alias{neisDa}
\title{Calculates Nei's ancestral/net genetic distance}
\usage{
neisDa(dxy, pi)
}
\arguments{
\item{dxy}{calculated by \code{neisDxy}}

\item{pi}{calculated by \code{neispi}}
}
\value{
A \code{dataframe} net genetic distance between two populaitons
}
\description{
Calculates Nei (1987) da from dxy and pi
}
\details{
Authours: Chris Ward
calculates net genetic distance between two populations
}
