% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neisDxy.R
\name{neisDxy}
\alias{neisDxy}
\title{Calculates Nei's Dxy}
\usage{
neisDxy(distMat, popList, pairs, ploidy)
}
\arguments{
\item{distMat}{hamming distance matrix calculated using the hidden function \code{genoDist}}

\item{popList}{List of populations made from \code{pops} dataframe provided by users}

\item{pairs}{list of populaiton pairs generated in \code{getDiversityStats}}

\item{ploidy}{\code{numeric} number of chromosomes}
}
\value{
A \code{dataframe} mean absolute genetic distance between samples
}
\description{
Calculates Nei (1987) Dxy from hamming distance
}
\details{
Authours: Chris Ward
calculates mean absolute genetic distance between two populations
}
