% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neisPi.R
\name{neisPi}
\alias{neisPi}
\title{Calculates Nei's nucleotide diversity (pi)}
\usage{
neisPi(distMat, popList, ploidy = 2, sampleCorrect = FALSE)
}
\arguments{
\item{distMat}{hamming distance matrix calculated using the hidden function \code{genoDist}}

\item{popList}{List of populations made from \code{pops} dataframe provided by users}

\item{ploidy}{\code{numeric} number of chromosomes}
}
\value{
A \code{dataframe} mean nucleotide diversity within each populaiton
}
\description{
Calculates Nei (1987) pi from hamming distance
}
\details{
Authours: Chris Ward
calculates within population nucleotide diversity
}
