% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureFasta.R
\name{outputLociFasta}
\alias{outputLociFasta}
\title{Index genome according to provided exons}
\usage{
outputLociFasta(GDS, loci, dir, pops, nCores = 1, ploidy = 2,
  alleles = "seperate", minSites = 0.1, removeIndels = TRUE,
  fasta = NULL)
}
\arguments{
\item{removeIndels}{removes indels, if false this will mess up the alignemtn in output fasta. concensus will also no longer work.1}

\item{fasta}{\code{DNAStringSet} the output of \code{Biostrings::readDNAStringSet()}. Input genome to extract features from, will align
automatically to the reference positions.}

\item{genome}{genome object of \code{DNAStringSet} class}

\item{exons}{GrangeList exons generated using \code{getFeatures} by passing \code{"gene:cds"} to \code{feature}}
}
\value{
fasta in specified directory.
}
\description{
Author: Christopher Ward \cr
Converts GDS directly to fasta and outputs to a directory. \cr
Providing a range to the function will result in multiple files (number of ranges) being output with alignemnts within the ranges.
}
\examples{

}
