% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputTrees.R
\name{outputTrees}
\alias{outputTrees}
\title{Output nj distance trees}
\usage{
outputTrees(GDS, loci, dir, pops, nCores = 1, ploidy = 2,
  alleles = "seperate", minSites = 0.1, removeIndels = TRUE)
}
\arguments{
\item{GDS}{genome object of \code{DNAStringSet} class}
}
\value{
fasta in specified directory.
}
\description{
Author: Christopher Ward \cr
Will export rough, distance based nj trees for each locus to a given directory.
}
\examples{

}
