% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evgram-print.R
\name{print.evgram}
\alias{print.evgram}
\title{Print \code{evgram} object}
\usage{
\method{print}{evgram}(x, ...)
}
\arguments{
\item{x}{An \code{evgram} object produced by the \code{\link[gear]{evgram}} function.}

\item{...}{Not currently implemented.}
}
\description{
Print an object of class \code{evgram} produced by the 
\code{\link[gear]{evgram}} function.
}
\examples{
data(co)
evgram(Al ~ 1, co, ~ easting + northing)
}
\author{
Joshua French
}
