% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDist.R
\name{relDxy}
\alias{relDxy}
\title{Calculates Relative genetic distance (dxy)}
\usage{
relDxy(distMat, popList, ploidy, outgroup)
}
\arguments{
\item{distMat}{hamming distance matrix calculated using the hidden function \code{genoDist}}

\item{popList}{List of populations made from \code{pops} dataframe provided by users}

\item{ploidy}{\code{numeric} number of chromosomes}

\item{outgroup}{\code{character} the furthest outgroup}
}
\value{
A \code{dataframe} containing Relative distance
}
\description{
Calculates Relative distance, an extension of ttd from Ward and Baxter 2018
}
\details{

}
