% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/federRND.R
\name{RND}
\alias{RND}
\title{Calculates Relative Node Depth}
\usage{
RND(distMat, popList, pairs, ploidy, outgroup, type = "feder")
}
\arguments{
\item{distMat}{hamming distance matrix calculated using the hidden function \code{genoDist}}

\item{popList}{List of populations made from \code{pops} dataframe provided by users}

\item{pairs}{list of populaiton pairs generated in \code{getDiversityStats}}

\item{ploidy}{\code{numeric} number of chromosomes}

\item{outgroup}{2}
}
\value{
A \code{dataframe} containing relative node depth
}
\description{
Calculates Relative Node Depth from Feder (2005)
}
\details{
Authours: Chris Ward
}
