% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpWindowMaker.R
\name{snpWindowMaker}
\alias{snpWindowMaker}
\alias{snpWindowMaker,data.frame-method}
\alias{snpWindowMaker,data_frame}
\title{Sliding or Tiled snpwindow maker.}
\usage{
snpWindowMaker(
  x,
  windowSize,
  stepSize,
  nCores = 1,
  GDS = NULL,
  snpWindow = FALSE,
  loci = NULL,
  ...
)

\S4method{snpWindowMaker}{data.frame}(
  x,
  windowSize,
  stepSize,
  nCores,
  GDS = NULL,
  snpWindow = FALSE,
  loci = NULL
)
}
\arguments{
\item{x}{Can be a \code{data.frame} object containing ID and Length columns for each scaffold, \cr
\code{Granges} or \code{GRangesList} object}

\item{stepSize}{\code{numeric} \cr Window step size in base pairs. Default is 0, specifies a tiled window.}

\item{WindowSize}{\code{numeric} \cr Window size in base pairs}
}
\value{
A \code{GRangesList} object where each element is a genomic window.
}
\description{
Makes sliding and tiled windows from data frame
}
\details{
Authours: Chris Ward & Alastair Ludington \cr
Generates tiled (no step size) and sliding (step size) ranges for analysis with other functions.
}
