% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeStats.R
\name{summarizeStats}
\alias{summarizeStats}
\title{Calculates block jackknifed mean for statistics}
\usage{
summarizeStats(x, blockSize = 1e+06, which = NULL, nCores = 1,
  stats = NULL)
}
\arguments{
\item{x}{A \code{matrix} \cr
Allele genotypes for each individual}

\item{blockSize}{\code{logical}
If \code{TRUE} missing data will be removed from distance calculations in a paiwise manner.}
}
\value{
A \code{matrix} of hamming distance between individuals
}
\description{
summarize statistics using a block jacknife
}
\details{
Authours: Chris Ward
}
