% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{A toy data set for examples}
\format{
A data frame with 25 rows and 3 columns:
\describe{
 \item{y}{response}
 \item{x1}{coordinate dimension 1 values}
 \item{x2}{coordinate dimension 2 values}
}
}
\usage{
data(toydata)
}
\description{
This is a toy data set of 25 observations 
generated on a [0, 1] x [0, 1] domain.
}
