% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgram.R
\name{vgram_check_args}
\alias{vgram_check_args}
\title{Check arguments of vgram}
\usage{
vgram_check_args(formula, data, coords, nbins, maxd, angle, ndir, type, npmin)
}
\description{
Check arguments of vgram
}
\keyword{internal}
