% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedFst.R
\name{weightedFst}
\alias{weightedFst}
\title{Calculates weighted Fst}
\usage{
weightedFst(distMat, popList, pairs, ploidy)
}
\arguments{
\item{distMat}{hamming distance matrix calculated using the hidden function \code{genoDist}}

\item{popList}{List of populations made from \code{pops} dataframe provided by users}

\item{pairs}{list of populaiton pairs generated in \code{getDiversityStats}}

\item{ploidy}{\code{numeric} number of chromosomes}
}
\value{
A \code{dataframe} Fst betweeb two populations
}
\description{
Fst from hamming distance
}
\details{
Authours: Chris Ward
calculates Fst between all populations
}
