% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowMaker-S4.R
\docType{methods}
\name{windowMaker}
\alias{windowMaker}
\alias{windowMaker,data.frame-method}
\alias{windowMaker,data_frame}
\alias{windowMaker,GRanges-method}
\alias{windowMaker,GRanges}
\alias{windowMaker,GRangesList-method}
\alias{windowMaker,GRangesList}
\title{Sliding or Tiled window maker.}
\usage{
windowMaker(x, windowSize, stepSize, ...)

\S4method{windowMaker}{data.frame}(x, windowSize, stepSize = 0,
  nCores = 1)

\S4method{windowMaker}{GRanges}(x, windowSize, stepSize = 0)

\S4method{windowMaker}{GRangesList}(x, windowSize, stepSize = 0,
  nCores = 1)
}
\arguments{
\item{x}{Can be a \code{data.frame} object containing ID and Length columns for each scaffold, \cr
\code{Granges} or \code{GRangesList} object}

\item{stepSize}{\code{numeric} \cr Window step size in base pairs. Default is 0, specifies a tiled window.}

\item{WindowSize}{\code{numeric} \cr Window size in base pairs}
}
\value{
A \code{GRangesList} object where each element is a genomic window.
}
\description{
Makes sliding and tiled windows from data frame
}
\details{
Authours: Chris Ward & Alastair Ludington \cr
Generates tiled (no step size) and sliding (step size) ranges for analysis with other functions.
}
\examples{

#### tiled windows ####
# make tiled windows of width 10,000 bp using 4 cores
windowSize <- 10000
stepSize <- 0
nCores <- 4

#### sliding windows ####
# make sliding windows of width 10,000 bp with 1000 bp step size,  using 4 cores
windowSize <- 10000
stepSize <- 1000
nCores <- 4


#### Using data.frame ####

# get contig metadata from fasta index
df <- readr::read_tsv("referene.fna.fai", col_names = FALSE)
x <- tibble::data_frame(ID = df[[1]], length = df[[2]])

# subset by desired contigs

x <- x[x$ID \%in\% c("chr1", "chr2"),]

# construct windows

windows <- windowMaker(x, windowSize, stepSize, nCores)

#### Using GRanges ####

# make GRange object of width 1 Mbp on chr1

x <- GRanges(seqnames = "chr1", ranges = IRanges(start = 1, end = 1000000))
windows <- windowMaker(x, windowSize, stepSize)

#### Using GRangesList ####

# make GRange object of width 1 Mbp on chr1 and chr2
x <-GRangesList(GRanges(seqnames = "chr1", ranges = IRanges(start = 1, end = 1000000)), GRanges(seqnames = "chr2", ranges = IRanges(start = 1, end = 1000000)))

windows <- windowMaker(x, windowSize, stepSize, nCores)

}
