% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowMaker.R
\name{windowMaker}
\alias{windowMaker}
\title{Sliding and tiled window ranges}
\usage{
windowMaker(x, windowSize, stepSize = 0, nCores = 1)
}
\arguments{
\item{stepSize}{\code{numeric} Window step size in base pairs. Default is 0, specifies a tiled window.}

\item{contigMD}{data frame with contig metadata. Contains two columns: \code{ID} with scaffold names, \code{length} with scaffold length}

\item{WindowSize}{\code{numeric} Window size in base pairs}
}
\value{
A \code{list} of \code{GRanges}
}
\description{
makes sliding and tiled windows from data frame
}
\details{
Authours: Chris Ward & Alastair Ludington
generates tiled (no step size) and sliding (step size) ranges for analysis with other functions.
}
\examples{

# Set up tiled windows
windowSize <- 100000
stepSize <- 0

# make contig metadata across all contigs in VCF header
#contigMD <- seqVCF_Header(vcf_path)$contig

# subset by desired contigs
#contigMD <- contigMD[contigMD$ID \%in\% c("chr1", "chr2"),]


}
