% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoDist.R
\name{genoDist}
\alias{genoDist}
\title{Calculates teh hamming distance between alleles}
\usage{
genoDist(genoMat)
}
\arguments{
\item{genoMat}{A \code{matrix} containing alleles for each individual}
}
\value{
A \code{matrix} of hamming distance between individuals
}
\description{
Used in the calculation of diversity statistics
}
\details{
Authours: Chris Ward
Calculates the hamming distance using matrix multiplication
}
