% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCodonIndex.R
\name{getCodonFeatures}
\alias{getCodonFeatures}
\title{Index genome according to provided exons}
\usage{
getCodonFeatures(genome, exons, nCores)
}
\arguments{
\item{genome}{genome}

\item{exons}{GrangeList exons generated using \code{getFeatures} by passing \code{"gene:exons"} to \code{feature}}
}
\value{
An object of \code{CodonIndex}
}
\description{
Author: Christopher Ward Index genome according to provided exons
}
\examples{

}
