% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDiversityStats.R
\docType{methods}
\name{getDiversityStats}
\alias{getDiversityStats}
\alias{getDiversityStats,character,ANY-method}
\alias{getDiversityStats,data_frame}
\alias{getDiversityStats,ANY,GRangesList-method}
\alias{getDiversityStats,SeqVarsGDSClass-GRangesList}
\alias{getDiversityStats,ANY,GRanges-method}
\alias{getDiversityStats,SeqVarsGDSClass-GRanges}
\title{Calculate diversity statistics}
\usage{
getDiversityStats(GDS, loci, minSites = 0.5, nCores = 1, pops = NULL,
  stats = "all", ploidy = 2, pairwiseDeletion, ...)

\S4method{getDiversityStats}{character,ANY}(GDS, loci, minSites = 0.5,
  nCores = 1, pops = NULL, stats = "all", ploidy = 2,
  pairwiseDeletion, ...)

\S4method{getDiversityStats}{ANY,GRangesList}(GDS, loci, minSites = 0.5,
  nCores = 1, pops = NULL, stats = "all", ploidy = 2,
  pairwiseDeletion)

\S4method{getDiversityStats}{ANY,GRanges}(GDS, loci, minSites = 0.5,
  nCores = 1, pops = NULL, stats = "all", ploidy = 2,
  pairwiseDeletion)
}
\arguments{
\item{GDS}{\code{GDS} object with variant data to import genotypes from.}

\item{loci}{\code{GRangesList}. Loci to import genotypes for.}

\item{minSites}{\code{numeric} minimum number of sites as a proportion of loci length. Default 0.5 ie 50 percent}

\item{nCores}{\code{numeric} number of cores to run in parallel}

\item{pops}{\code{data_frame} containing Sample ID and poplations. Default is to treat individuals as populations \code{c("none")}}

\item{stats}{\code{character}. Vector containing all diversity stats to calculate. default \code{c("all")}}

\item{ploidy}{\code{numeric} number of alleles}

\item{pairwiseDeletion}{\code{numeric} nshould Ns be removed from pairwise comparisons}
}
\value{
A \code{data_frame} of selected Diversity statistics
}
\description{
Imports genotypes and calculates diversity formula
}
\details{
Authour: Chris Ward
Uses a \code{list} containing \code{GRanges} or a \code{GRangesList} to import genotypes. Diversity statistics are then calculated.
}
