% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgram.R
\name{omni_semivariogram}
\alias{omni_semivariogram}
\title{determine omnidireciton semivariogram based on 
based on distances, kernel (the numerator of the estimator),
the cuts (tolerance regions), and minimum numer of pairs allowed}
\usage{
omni_semivariogram(d, diff, cuts, npmin, type)
}
\description{
determine omnidireciton semivariogram based on 
based on distances, kernel (the numerator of the estimator),
the cuts (tolerance regions), and minimum numer of pairs allowed
}
\keyword{internal}
