% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsummaryStats.R, R/summarizeStats.R
\name{summarizeStats}
\alias{summarizeStats}
\title{Calculates block jackknifed mean for statistics}
\usage{
summarizeStats(
  x,
  blockSize = 1000000L,
  which = NULL,
  nCores = 1,
  stat = "diversity"
)

summarizeStats(
  x,
  blockSize = 1000000L,
  which = NULL,
  nCores = 1,
  stat = "diversity"
)
}
\arguments{
\item{x}{A \code{gear} \cr
Analyzed gear object}

\item{blockSize}{\code{integer} default \code{1000000L}. The block size in
basepairs.}

\item{which}{\code{character,NULL} default \code{NULL}. Only calculates
across specified scaffolds/chromosomes. if  \code{NULL} will calculate
across all windows provided.}

\item{nCores}{\code{numeric} Number of cores allcated for calculations.}

\item{stats}{\code{character,NULL} default \code{NULL}. Only calculates
for specified statistics contained in the dataframe. if  \code{NULL} will
calculate across all statistics provided.}
}
\value{
A \code{dataframe} with mean, jackSE, Zscore, median, and maximum for
each statistic.

A \code{dataframe} with mean, jackSE, Zscore, median, and maximum for
each statistic.
}
\description{
summarize statistics using a block jacknife

summarize statistics using a block jacknife
}
\details{
Authours: Chris Ward

Authours: Chris Ward
}
