% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate4FoldCodons.R
\name{validate4FoldCodons}
\alias{validate4FoldCodons}
\title{Validate 4fold degenerate codons}
\usage{
validate4FoldCodons(GDS, input, nCores, pops)
}
\value{
An object of class \code{GRanges} containing codon ranges for positions specified.
}
\description{
This function will validate 4 fold degenerate codons using a provided SNP dataset
You only need to run this if you are interested in 4 fold degenerate sights 
Usually used for their quasi neutrality
}
\examples{

}
